Refactor: Extract order status update logic
Moved order status update logic to a separate hook for better organization and maintainability.  Added email functionality.
