+++ src/collections/Orders/index.ts
@@ -1,0 +1,42 @@
+import type { CollectionConfig } from 'payload'
+import { afterChangeStatus } from './hooks/afterChangeStatus'
+
+export const Orders: CollectionConfig = {
+  slug: 'orders',
+  admin: {
+    useAsTitle: 'orderCode',
+  },
+  fields: [
+    {
+      name: 'orderCode',
+      type: 'text',
+    },
+    {
+      name: 'user',
+      type: 'relationship',
+      relationTo: 'users',
+    },
+    {
+      name: 'status',
+      type: 'select',
+      defaultValue: 'processing',
+      options: [
+        { label: 'Processing', value: 'processing' },
+        { label: 'Canceled', value: 'canceled' },
+        { label: 'Completed', value: 'completed' },
+        { label: 'Failed', value: 'failed' },
+      ],
+      hooks: {
+        afterChange: [afterChangeStatus],
+      },
+    },
+    {
+      name: 'total',
+      type: 'number',
+    },
+    {
+      name: 'currency',
+      type: 'text',
+    },
+  ],
+}
