+++ src/collections/Tickets/index.ts
@@ -1,5 +1,5 @@
 import type { CollectionConfig } from 'payload'
-import { afterChangeStatus } from './hooks/afterChangeStatus'
+// import { afterChangeStatus } from './hooks/afterChangeStatus'
 
 export const Tickets: CollectionConfig = {
   slug: 'tickets',
@@ -63,7 +63,7 @@ export const Tickets: CollectionConfig = {
       ],
       required: false,
       hooks: {
-        afterChange: [afterChangeStatus],
+        // afterChange: [afterChangeStatus],
       },
     },
   ],
